% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotCovariates.R
\name{barplotCovariates}
\alias{barplotCovariates}
\title{Barplot Covariates}
\usage{
barplotCovariates(df)
}
\arguments{
\item{df}{an object of class 'data.frame'}
}
\description{
Create barplot of injury covariates from data frame column 'Narrative',
stratified by known health status ("DEAD.DECLINE" vs "RECOVERED")
}
\examples{
barplotCovariates(df)

}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
