% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateStrings.R
\name{CovariateStrings}
\alias{CovariateStrings}
\title{CovariateStrings}
\usage{
CovariateStrings(df)
}
\arguments{
\item{df}{an object of class 'data.frame'}
}
\description{
Use existing randomForest entanglement and vessel strike models to output data frame
that includes health status assignment probabilities ('DEAD.DECLINE' vs 'RECOVERED'),
and covariate states / presence from the 'Narrative' field.
}
\examples{

# Show model predictions alongside covariate states

CovList <- CovariateStrings(data.test.entangle)
ENprobs <- predict(ModelEntangle, data.test.entangle, type="prob")
EN.df.probs <- cbind.data.frame(CovList, ENprobs)
head(EN.df.probs)

}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
