% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Numeric_From_String.R
\name{Extract_Numeric_From_String}
\alias{Extract_Numeric_From_String}
\title{Extract_Numeric_From_String}
\usage{
Extract_Numeric_From_String(df)
}
\arguments{
\item{df}{an object of class 'data.frame'}
}
\description{
Extract numeric values appearing in data frame narratives to be used for VessSpd and VessSz covariate definitions
}
\examples{

# Return a vector of unique numeric values appearing in object ʻdfʻ

Values <- Extract_Numeric_From_String(WhaleData)


}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
