% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monofilament_Hook_Line.R
\name{Monofilament_Hook_Line}
\alias{Monofilament_Hook_Line}
\title{Monofilament_Hook_Line}
\usage{
Monofilament(df)
}
\arguments{
\item{df}{an object of class 'data.frame'}
}
\description{
Create a covariate to identify monofilament line, but exclude monofilament gillnets.
Targeted toward hook and line entanglements involving monofilament line.
}
\examples{
# Identify which cases include non-gillnet monofilament

Monofilament <- Monofilament(WhaleData)

}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
