% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Run Shiny interface with SeriousInjury}
\usage{
run_shiny()
}
\value{
None
}
\description{
\code{run_shiny} runs a shiny application for the main functions in SeriousInjury.
}
\details{
Note:
}
