% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFSeriousInjury.R
\name{SeriousInjury}
\alias{SeriousInjury}
\title{SeriousInjury()}
\usage{
SeriousInjury()
}
\description{
Random Forest (RF) Method For Assessment of Whale Injuries
}
\examples{
SeriousInjury(25)

}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
