% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_Injury_Models.R
\name{Update_Injury_Models}
\alias{Update_Injury_Models}
\title{Update_Injury_Models}
\usage{
Update_Injury_Models()
}
\description{
Generate Random Forest (RF) entanglement and vessel strike injury models.

Note that this function overwrites the "WhaleData.RData" file
in both 'data' and 'extdata'
}
\examples{
ModelEntangle
ModelVessel
data.entangle
data.vessel

}
