% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VesselSpeed.R
\name{VessSpd}
\alias{VessSpd}
\title{VessSpd}
\usage{
VessSpd(df)
}
\arguments{
\item{df}{an object of class 'data.frame'}
}
\description{
Assess whether narratives include reference to vessel speed.
Classify vessel speeds as unknown, slow, or fast, based on a threshold of <=10 KTs being ʻslowʻ
}
\examples{

# append covariate VessSpd to WhaleData

VessSpd <- VessSpd(WhaleData)


}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
