% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VesselSize.R
\name{VessSz}
\alias{VessSz}
\title{VessSz}
\usage{
VessSz(df)
}
\arguments{
\item{df}{an object of class 'data.frame'}
}
\description{
Assess whether narratives include reference to vessel size.
Classify vessel sizes as small, large or unknown, using >=65 FT as threshold for large vessels
}
\examples{

# append covariate VessSz to WhaleData

VessSz <- VessSz(WhaleData)


}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
