% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhaleData.R
\docType{data}
\name{WhaleData}
\alias{WhaleData}
\title{WhaleData}
\format{
WhaleData an object of class 'data.frame'
}
\usage{
data(WhaleData)
}
\description{
Whale injury data for package \emph{SeriousInjury}. Data frame of > 1000 whale injury cases.
'WhaleData' includes field 'Narrative' (case-sensitive) used to generate injury covariates
for use with Random Forest models used to predict health status.
}
\examples{
data(WhaleData)
head(WhaleData)


}
\keyword{datasets}
