% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhaleInjuryCovariates.R
\name{WhaleInjuryCovariates}
\alias{WhaleInjuryCovariates}
\title{WhaleInjuryCovariates()}
\usage{
WhaleInjuryCovariates(x)
}
\arguments{
\item{x}{an object of class 'data.frame'}
}
\description{
Identify Covariates for Whale Injury Assessments from
Narratives + Append Covariates to existing data.frame.

Search wide-form data.frame column named 'Narrative' for words/phrases
to be coded as binary absence / presence (0/1) injury covariates.
Append to new data.frame with function WhaleInjuryCovariates().

Resulting data.frame is used with randomForest models to predict
health status of whales involved in entanglements or vessel strikes.

Package includes example data.frame "WhaleData"
}
\examples{
WhaleInjuryCovariates(WhaleData)

}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
