% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhaleInjuryCovariates.R
\name{WhaleInjuryCovariates}
\alias{WhaleInjuryCovariates}
\title{WhaleInjuryCovariates}
\usage{
WhaleInjuryCovariates(x)
}
\arguments{
\item{x}{an object of class 'data.frame'}
}
\description{
Define covariates from whale injury narratives based on presence or absence of key words or phrases
used to predict health outcomes. 'Narratives' (case sensitive) should be in a wide-form data frame
in column of same name. Append covariates to data frame using function WhaleInjuryCovariates(). Resulting
data frame is used with Random Forest classification trees to predict health status of whales involved in
entanglements or vessel strikes. Package includes example data frame "WhaleData"
}
\examples{
WhaleInjuryCovariates(WhaleData)

}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
